# optimize topk for greedysearch