--- a/onnxruntime/contrib_ops/cpu/transformers/greedy_search_impl_base.h
+++ b/onnxruntime/contrib_ops/cpu/transformers/greedy_search_impl_base.h
@@ -36,7 +36,7 @@ struct SamplingState : public ISamplingState<T> {
       this->d_softmaxed_score = AllocateBuffer<float>(allocator, d_softmaxed_score_buffer_, SafeInt<size_t>(total_count));
       this->d_sampled = AllocateBuffer<float>(allocator, d_sampled_buffer_, SafeInt<size_t>(batch_size));
       this->h_sampled_all = AllocateBuffer<float>(cpu_allocator, h_sampled_all_buffer_, SafeInt<size_t>(batch_size * max_iter));
-      this->d_indices = AllocateBuffer<int64_t>(allocator, d_indices_buffer_, SafeInt<size_t>(batch_size));
+      this->d_indices = AllocateBuffer<int32_t>(allocator, d_indices_buffer_, SafeInt<size_t>(batch_size));
       this->temp_storage_bytes = 0;
       // TODO: Do not allocate this buffer if there's no presence_mask
       this->d_presence_mask = AllocateBuffer<int>(allocator, d_presence_mask_buffer_, SafeInt<size_t>(total_count));
@@ -79,7 +79,7 @@ struct GreedySearchState : public IGreedySearchState<T> {
             int vocab_size,
             int sequence_length,
             int max_length,
-            bool /*is_cuda*/) {
+            bool is_cuda) {
     // below buffers are on cpu
     this->sequences_space = AllocateBuffer<int32_t>(cpu_allocator,
                                                     sequences_space_buffer_,
@@ -91,15 +91,23 @@ struct GreedySearchState : public IGreedySearchState<T> {
     this->eos_meet = AllocateBuffer<bool>(cpu_allocator, eos_meet_buffer_, batch_size);
     memset(this->eos_meet.data(), 0, this->eos_meet.size_bytes());
 
-    this->next_tokens_cpu = AllocateBuffer<int64_t>(cpu_allocator,
-                                                    next_tokens_cpu_buffer_,
-                                                    SafeInt<size_t>(batch_size));
     this->next_tokens = AllocateBuffer<int32_t>(cpu_allocator, next_tokens_buffer_, SafeInt<size_t>(batch_size));
 
     // below buffers are on cpu or cuda
     size_t next_token_size = SafeInt<size_t>(batch_size) * vocab_size;
     this->next_token_scores = AllocateBuffer<T>(allocator, next_token_scores_buffer_, next_token_size);
     this->next_positions = AllocateBuffer<int32_t>(allocator, next_positions_buffer_, batch_size);
+
+    if (is_cuda) {
+      AllocateTempBufferForGetGreedySearchTopOne<T>(
+          batch_size,
+          allocator,
+          this->temp_topk_buffer_,
+          this->temp_topk_scores_buffer,
+          this->temp_topk_tokens_buffer,
+          this->topk_scores_buffer,
+          this->topk_tokens_buffer);
+    }
   }
 
   void SetSequence(gsl::span<const int32_t> input_ids_in_cpu,
@@ -109,8 +117,8 @@ struct GreedySearchState : public IGreedySearchState<T> {
     gsl::span<int32_t> sequences_0 = this->sequences_space;
     for (size_t i = 0; i < batch_beam_size; i++) {
       for (int j = 0; j < sequence_length; j++) {
-        sequences_0[SafeInt<gsl::index>(i) * max_length + j] = \
-        static_cast<int32_t>(input_ids_in_cpu[SafeInt<gsl::index>(i) * sequence_length + j]);
+        sequences_0[SafeInt<gsl::index>(i) * max_length + j] =
+            static_cast<int32_t>(input_ids_in_cpu[SafeInt<gsl::index>(i) * sequence_length + j]);
       }
     }
   }
@@ -120,9 +128,9 @@ struct GreedySearchState : public IGreedySearchState<T> {
   BufferUniquePtr sequence_lengths_buffer_;
   BufferUniquePtr next_token_scores_buffer_;
   BufferUniquePtr next_tokens_buffer_;
-  BufferUniquePtr next_tokens_cpu_buffer_;
   BufferUniquePtr next_positions_buffer_;
   BufferUniquePtr eos_meet_buffer_;
+  BufferUniquePtr temp_topk_buffer_;
 };
 
 // Base class of gready search implementation that is common for both GPT-2 and Bart/T5.
@@ -138,13 +146,13 @@ class GreedySearchBase : public GenerateBase {
                    const GenerationDeviceHelper::TopkFunc& topk_func,
                    const GenerationDeviceHelper::GreedySearchProcessLogitsFunc<T>& process_logits_func,
                    const GenerationDeviceHelper::DeviceCopyFunc<float>& device_copy_func)
-    : GenerateBase(context,
-                   decoder_session_state,
-                   thread_pool,
-                   ort_stream,
-                   cuda_dumper,
-                   topk_func,
-                   device_copy_func),
+      : GenerateBase(context,
+                     decoder_session_state,
+                     thread_pool,
+                     ort_stream,
+                     cuda_dumper,
+                     topk_func,
+                     device_copy_func),
         parameters_(&params),
         process_logits_func_(process_logits_func) {
     parameters_->ParseFromInputs(&context);
@@ -186,11 +194,11 @@ Status GreedySearchBase<T, ParametersT>::CheckInputs(const OpKernelContextIntern
   //   input_ids  : (batch_size, sequence_length)
   //   vocab_mask : (vocab_size) or nullptr
   ORT_RETURN_IF_ERROR(this->CheckInputsImpl(parameters_,
-                                            context.Input<Tensor>(0),   // input_ids
-                                            context.Input<Tensor>(4),   // vocab_mask
-                                            context.Input<Tensor>(5),   // prefix_vocab_mask
-                                            context.Input<Tensor>(6),   // attention_mask
-                                            context.Input<Tensor>(7))); // presence_mask
+                                            context.Input<Tensor>(0),    // input_ids
+                                            context.Input<Tensor>(4),    // vocab_mask
+                                            context.Input<Tensor>(5),    // prefix_vocab_mask
+                                            context.Input<Tensor>(6),    // attention_mask
+                                            context.Input<Tensor>(7)));  // presence_mask
 
   return Status::OK();
 }
@@ -241,10 +249,6 @@ Status GreedySearchBase<T, ParametersT>::GenerateNextToken(
   ORT_RETURN_IF_ERROR(ProcessLogits(logits, greedy_state, sampling_state, this->temp_space_allocator_, counter));
 
   next_tokens = greedy_state.next_tokens;
-  for (size_t i = 0; i < next_tokens.size(); i++) {
-    next_tokens[i] = gsl::narrow_cast<int32_t>(greedy_state.next_tokens_cpu[i]);
-  }
-
   gsl::span<bool>& eos_meet = greedy_state.eos_meet;
   for (size_t batch_id = 0; batch_id < next_tokens.size(); ++batch_id) {
     if (next_tokens[batch_id] == eos_token_id || eos_meet[batch_id] == true) {