--- a/onnxruntime/contrib_ops/cuda/transformers/generation_device_helper.cc
+++ b/onnxruntime/contrib_ops/cuda/transformers/generation_device_helper.cc
@@ -16,6 +16,7 @@
 #include "contrib_ops/cpu/transformers/subgraph_t5_decoder.h"
 #include "contrib_ops/cpu/transformers/subgraph_gpt.h"
 #include "contrib_ops/cuda/transformers/beam_search_topk.h"
+#include "contrib_ops/cuda/transformers/greedy_search_top_one.h"
 #include "core/providers/cuda/nvtx_profile.h"
 #include "core/providers/cuda/nvtx_profile_context.h"
 #include "sampling_cuda_helper.h"
@@ -584,7 +585,7 @@ Status GreedySearchProcessLogits(
 #ifdef DEBUG_GENERATION
   dumper->Print("logits", logits);
   if (is_reuse_logits_buffer) {
-    //TODO: Handle padded logits in the logits buffer before printing its contents
+    // TODO: Handle padded logits in the logits buffer before printing its contents
     ORT_THROW("Dumping contents of logits buffer is not implemented yet");
   } else {
     dumper->Print("next_token_scores", next_token_scores.data(), batch_size, vocab_size);
@@ -639,7 +640,7 @@ Status GreedySearchProcessLogits(
 
 #ifdef DEBUG_GENERATION
   if (is_reuse_logits_buffer) {
-    //TODO: Handle padded logits in the logits buffer before printing its contents
+    // TODO: Handle padded logits in the logits buffer before printing its contents
     ORT_THROW("Dumping contents of logits buffer is not implemented yet");
   } else {
     dumper->Print("next_token_scores after logits process", next_token_scores.data(), batch_size, vocab_size);
@@ -662,43 +663,26 @@ Status GreedySearchProcessLogits(
     return Status::OK();
   }
 
-  // next_tokens = torch.argmax(scores, dim=-1)
-  int64_t next_token_scores_dims[] = {static_cast<int64_t>(batch_size),
-                                      is_reuse_logits_buffer ? padded_vocab_size : vocab_size};
-  TensorShape next_token_scores_shape(&next_token_scores_dims[0], 2);
-  auto element_type = DataTypeImpl::GetType<T>();
-  OrtValue next_token_scores_value;
-
-  // TODO(hasesh): Same TODO as above about avoiding the const_cast here
-  Tensor::InitOrtValue(element_type,
-                       next_token_scores_shape,
-                       is_reuse_logits_buffer
-                           ? const_cast<void*>(reinterpret_cast<const void*>(logits_data))
-                           : reinterpret_cast<void*>(next_token_scores.data()),
-                       allocator->Info(),
-                       next_token_scores_value);
-  const Tensor& input = next_token_scores_value.Get<Tensor>();
-
-  constexpr int axis = 1;
-  constexpr unsigned top_k = static_cast<unsigned>(1);
-  constexpr bool largest = true;
-  constexpr bool sorted = false;
-
-  auto topk_scores = Tensor::CreateDefault();
-  auto topk_indices = Tensor::CreateDefault();
-  ORT_RETURN_IF_ERROR(TopK(&input, axis, top_k, largest, sorted, allocator, stream, thread_pool,
-                           *topk_scores, *topk_indices));
+  const CudaT* top_one_input = is_reuse_logits_buffer ? logits_data
+                                                      : reinterpret_cast<const CudaT*>(next_token_scores.data());
+  cuda::GreedySearchTopOne(
+      top_one_input,
+      batch_size,
+      is_reuse_logits_buffer ? padded_vocab_size : vocab_size,
+      reinterpret_cast<CudaT*>(greedy_state->temp_topk_scores_buffer.data()),
+      greedy_state->temp_topk_tokens_buffer.data(),
+      reinterpret_cast<CudaT*>(greedy_state->topk_scores_buffer.data()),
+      greedy_state->topk_tokens_buffer.data(),
+      cuda_stream);
 
 #ifdef DEBUG_GENERATION
-  dumper->Print("topk_scores", *(topk_scores.get()));
-  dumper->Print("topk_indices", *(topk_indices.get()));
+  dumper->Print("topk_scores", greedy_state->topk_scores_buffer.data(), batch_size);
+  dumper->Print("topk_indices", greedy_state->topk_tokens_buffer.data(), batch_size);
 #endif
 
-  const int64_t* next_token_indices = topk_indices->Data<int64_t>();
-
-  CUDA_RETURN_IF_ERROR(cudaMemcpyAsync(greedy_state->next_tokens_cpu.data(),
-                                       next_token_indices,
-                                       greedy_state->next_tokens_cpu.size_bytes(),
+  CUDA_RETURN_IF_ERROR(cudaMemcpyAsync(greedy_state->next_tokens.data(),
+                                       greedy_state->topk_tokens_buffer.data(),
+                                       greedy_state->next_tokens.size_bytes(),
                                        cudaMemcpyDeviceToHost,
                                        cuda_stream));
   CUDA_RETURN_IF_ERROR(cudaStreamSynchronize(cuda_stream));