--- a/onnxruntime/contrib_ops/cuda/transformers/generation_device_helper.cc
+++ b/onnxruntime/contrib_ops/cuda/transformers/generation_device_helper.cc
@@ -526,6 +526,7 @@ Status GreedySearchProcessLogits(
 #endif
 
   ORT_UNUSED_PARAMETER(logits_processors);
+  ORT_UNUSED_PARAMETER(thread_pool);
 #ifndef DEBUG_GENERATION
   ORT_UNUSED_PARAMETER(dumper);
 #endif