--- a/onnxruntime/contrib_ops/cuda/transformers/sampling_cuda_helper.h
+++ b/onnxruntime/contrib_ops/cuda/transformers/sampling_cuda_helper.h
@@ -150,7 +150,7 @@ Status Sample(AllocatorPtr& allocator,
   dumper->Print("d_sampled", d_sampled.data(), parameters->batch_size, 1);
 #endif
 
-  gsl::span<int64_t>& d_indices = sampling_state->d_indices;
+  gsl::span<int32_t>& d_indices = sampling_state->d_indices;
   gsl::span<int>& presence_mask = sampling_state->d_presence_mask;
   cuda::TorchMultinomialKernelLauncher(d_softmaxed_score.data(),
                                        d_sampled.data(),
@@ -164,9 +164,9 @@ Status Sample(AllocatorPtr& allocator,
   dumper->Print("d_indices", d_indices.data(), parameters->batch_size, 1);
 #endif
 
-  CUDA_RETURN_IF_ERROR(cudaMemcpyAsync(greedy_state->next_tokens_cpu.data(),
+  CUDA_RETURN_IF_ERROR(cudaMemcpyAsync(greedy_state->next_tokens.data(),
                                        sampling_state->d_indices.data(),
-                                       greedy_state->next_tokens_cpu.size_bytes(),
+                                       greedy_state->next_tokens.size_bytes(),
                                        cudaMemcpyDeviceToHost,
                                        cuda_stream));