--- a/onnxruntime/core/providers/cuda/test/all_tests.h
+++ b/onnxruntime/core/providers/cuda/test/all_tests.h
@@ -9,6 +9,7 @@ namespace test {
 bool TestDeferredRelease();
 bool TestDeferredReleaseWithoutArena();
 bool TestBeamSearchTopK();
+bool TestGreedySearchTopOne();
 
 }  // namespace test
 }  // namespace cuda