--- a/onnxruntime/core/providers/cuda/test/beam_search_topk.cc
+++ b/onnxruntime/core/providers/cuda/test/beam_search_topk.cc
@@ -127,6 +127,8 @@ bool TestBeamSearchTopK() {
     }
   }
 
+  CUDA_CALL_THROW(cudaFree(cuda_buffer));
+
   return true;
 }