… onnxruntime/core/providers/cuda/test/greedy_search_top_one.cc